Software/requirements.txt
numpy
pyserial
matplotlib
opencv-python
